src/common/requirements.txt
boto3==1.28.9
django-environ==0.11.2
